Resizing/regridding speed and accuracy improvements. Issues with Delta MPI.

+++ make.sh
@@ -240,8 +240,10 @@ if [[ "$ARGS" == *"hdf5"* && "$ARGS" == *"clean"* ]]; then
       HDF_EXTRA="--enable-parallel"
     fi
   fi
-  CC=$HDF_CC sh configure -C $HDF_EXTRA --prefix=$PWD/../hdf5 --enable-build-mode=production \
+set -x
+  CC=$HDF_CC sh configure -C $HDF_EXTRA --prefix=$SOURCE_DIR/external/hdf5 --enable-build-mode=production \
   --disable-dependency-tracking --disable-hl --disable-tests --disable-tools --disable-shared --disable-deprecated-symbols
+set +x
   wait 1
 
   # Compiling C takes less memory
@@ -255,7 +257,7 @@ if [[ "$ARGS" == *"hdf5"* && "$ARGS" == *"clean"* ]]; then
   cd ../..
 fi
 if [[ "$ARGS" == *"hdf5"* ]]; then
-  PREFIX_PATH="$PWD/external/hdf5;$PREFIX_PATH"
+  PREFIX_PATH="$SOURCE_DIR/external/hdf5;$PREFIX_PATH"
 fi
 
 ### Build KHARMA ###
@@ -293,6 +295,7 @@ if [[ "$ARGS" == *"clean"* ]]; then
   fi
 fi
 
-make -j$NPROC
-
-cp kharma/kharma.* ..
+if [[ "$ARGS" != *"dryrun"* ]]; then
+  make -j$NPROC
+  cp kharma/kharma.* ..
+fi
